+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/util/ResourcePaths.java
@@ -4,30 +4,30 @@ import java.util.Collections;
 import java.util.List;
 
 public interface ResourcePaths {
-    ResourcePath<Object> MATCH_ALL = new ResourcePath<Object>() {
+    ResourcePath MATCH_ALL = new ResourcePath() {
         @Override
-        public ResourcePath<Object> combine(ResourcePath<Object> other) {
+        public ResourcePath combine(ResourcePath other) {
             return other;
         }
         @Override
         public List<String> getPatterns() { return Collections.emptyList(); }
         @Override
-        public boolean matches(Object path) {
+        public boolean matches(String path) {
             return true;
         }
         @Override public String toString() {
             return "ResourcePath$MATCH_ALL";
         }
     };
-    ResourcePath<Object> MATCH_NONE = new ResourcePath<Object>() {
+    ResourcePath MATCH_NONE = new ResourcePath() {
         @Override
-        public ResourcePath<Object> combine(ResourcePath<Object> other) {
+        public ResourcePath combine(ResourcePath other) {
             return other;
         }
         @Override
         public List<String> getPatterns() { return Collections.emptyList(); }
         @Override
-        public boolean matches(Object path) {
+        public boolean matches(String path) {
             return false;
         }
         @Override public String toString() {
@@ -35,13 +35,11 @@ public interface ResourcePaths {
         }
     };
 
-    @SuppressWarnings("unchecked")
-    static <T> ResourcePath<T> matchALL() {
-        return (ResourcePath<T>) MATCH_ALL;
+    static ResourcePath matchALL() {
+        return MATCH_ALL;
     }
 
-    @SuppressWarnings("unchecked")
-    static <T> ResourcePath<T> matchNone() {
-        return (ResourcePath<T>) MATCH_NONE;
+    static ResourcePath matchNone() {
+        return MATCH_NONE;
     }
 }